+++ src/sll/api/_generated.c
@@ -3808,6 +3808,47 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_string_reverse_raw(sll_o
 
 
 
+__SLL_API_TYPE_sll_api_string_secure_equal sll_api_string_secure_equal(__SLL_API_ARGS_sll_api_string_secure_equal);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_string_secure_equal_raw(sll_object_t*const* al,sll_arg_count_t all){
+	sll_object_t* a=NULL;
+	if (all>0){
+		a=*(al+0);
+		if (SLL_OBJECT_GET_TYPE(a)==SLL_OBJECT_TYPE_STRING){
+			SLL_ACQUIRE(a);
+		}
+		else{
+			a=sll_operator_cast(a,sll_static_int[SLL_OBJECT_TYPE_STRING]);
+		}
+	}
+	else{
+		a=SLL_CREATE();
+		a->t=SLL_OBJECT_TYPE_STRING;
+		sll_string_create(0,&(a->dt.s));
+	}
+	sll_object_t* b=NULL;
+	if (all>1){
+		b=*(al+1);
+		if (SLL_OBJECT_GET_TYPE(b)==SLL_OBJECT_TYPE_STRING){
+			SLL_ACQUIRE(b);
+		}
+		else{
+			b=sll_operator_cast(b,sll_static_int[SLL_OBJECT_TYPE_STRING]);
+		}
+	}
+	else{
+		b=SLL_CREATE();
+		b->t=SLL_OBJECT_TYPE_STRING;
+		sll_string_create(0,&(b->dt.s));
+	}
+	sll_bool_t out=sll_api_string_secure_equal(&(a->dt.s),&(b->dt.s));
+	SLL_RELEASE(b);
+	SLL_RELEASE(a);
+	SLL_ACQUIRE(sll_static_int[out]);
+	return sll_static_int[out];
+}
+
+
+
 __SLL_API_TYPE_sll_api_string_split sll_api_string_split(__SLL_API_ARGS_sll_api_string_split);
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_string_split_raw(sll_object_t*const* al,sll_arg_count_t all){
 	sll_object_t* a=NULL;
@@ -4776,6 +4817,11 @@ static const internal_function_t _ifunc_data_ptr[]={
 		SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL
 	},
 	{
+		"sll:string_secure_equal",
+		sll_api_string_secure_equal_raw,
+		SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL
+	},
+	{
 		"sll:string_split",
 		sll_api_string_split_raw,
 		SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL
@@ -4874,5 +4920,5 @@ static const internal_function_t _ifunc_data_ptr[]={
 
 
 
-const sll_function_index_t _ifunc_size=129;
+const sll_function_index_t _ifunc_size=130;
 const internal_function_t* _ifunc_data=(const internal_function_t*)(&_ifunc_data_ptr);